backend/requirements.txt
requests
psycopg2-binary
sentence-transformers
pdf2image
Pillow
